fixing setup script. adding tags on managed machines

+++ centos_auto_setup_machine.sh
@@ -1,11 +1,32 @@
 
 #! /bin/bash
+
+
+PROD_CONF_FILE=~/prod.conf
+PEM_FILE=~/hpcloud.pem
+SYSCONF_FILE=~/sysconfig_play
+
+if [ ! -f ${PROD_CONF_FILE} ]; then
+    echo "missing ${PROD_CONF_FILE}"
+    exit 1
+fi
+if [ ! -f ${PEM_FILE} ]; then
+    echo "missing ${PEM_FILE}"
+    exit 1
+fi
+
+if [ ! -f ${SYSCONF_FILE} ]; then
+    echo "missing ${SYSCONF_FILE}"
+    exit 1
+fi
+
+
 echo "installing java"
 yum  -y install java-1.6.0-openjdk-devel
 export JAVA_HOME=/usr/lib/jvm/java-1.6.0-openjdk.x86_64
 
 echo "downloading play"
-if [ -e play-2.0.4.zip ]; then
+if [ -f play-2.0.4.zip ]; then
     echo "play already exists - nothing to do"
 else
     wget 'http://download.playframework.org/releases/play-2.0.4.zip'
@@ -15,7 +36,7 @@ fi
 echo "downloading cloudify"
 CLOUDIFY_FOLDER=gigaspaces-cloudify-2.3.0-ga-b3510
 CLOUDIFY_FILE=${CLOUDIFY_FOLDER}.zip
-if [ -e $CLOUDIFY_FILE ]; then
+if [ -f $CLOUDIFY_FILE ]; then
     echo "cloudify already installed, nothing to go"
 else
     wget 'http://repository.cloudifysource.org/org/cloudifysource/2.3.0-RELEASE/${CLOUDIFY_FILE}'
@@ -37,17 +58,17 @@ gem install sass
 
 # assuming sysconfig_play exists on machine
 echo "copying sysconfig file"
-\cp -f ~/sysconfig_play /etc/sysconfig/play
+\cp -f ${SYSCONF_FILE} /etc/sysconfig/play
 . /etc/sysconfig/play
 
 # assuming there is a prod.conf copied to here
 echo "copying configuration files"
-\cp -f ~/prod.conf cloudify-widget/conf
-\cp -f ~/hpcloud.pem cloudify-widget/bin
-ln -fs ~/$CLOUDIFY_FILE bin/cloudify-folder # create a symbolic link to cloudify home.
+\cp -f ${PROD_CONF_FILE} cloudify-widget/conf
+\cp -f ${PEM_FILE} cloudify-widget/bin
+ln -fs ~/${CLOUDIFY_FILE} bin/cloudify-folder # create a symbolic link to cloudify home.
 chmod 755 cloudify-widget/*.sh
 chmod 755 cloudify-widget/bin/*.sh
-ln -s /root/$CLOUDIFY_FOLDER cloudify-widget/cloudify-folder
+ln -s /root/${CLOUDIFY_FOLDER} cloudify-widget/cloudify-folder
 
 #install mysql
 echo "installing mysql"
